Optimize ReasonReactRouter url comparison
cc @nikgraf @rickyvetter

As much as possible, we try not to drag in runtime in ReasonReact (and the current RR's `React.re` is runtime-free, apart from a few empty module outputs). This is not just to give a very comfortable and reliable guarantee of "zero-cost perf", but also to ensure there's no semantic mismatch between React.js and ReasonReact. As seen with... well, most codebases, any helper that inserts itself into two layers potentially breaks things, and old versions of RR certainly had bugs regarding discrepancies vs React.js. The relatively new hook-based version can guarantee "zero interop bug", which is something extremely nice to rely on whenever we can.

This is just a tangent though; but yeah, no runtime is great for many reasons. In this particular diff, the generic object comparison was added in 9147410fb0ee296950b8d79e9cd59eb392a7ee50. This isn't great:
- The generic comparison is theoretically wrong because it doesn't properly handle some of JS' corner-cases (imagine someone puts a `Symbol` on an object, etc.): https://github.com/BuckleScript/bucklescript/blob/80297bbb138f5e0d69c7fed74c5b6043391b2505/lib/js/caml_obj.js#L305
- The above code, because of its genericity, is also extremely slow, since it introspects into some runtime types, allocates (in a comparison!...), is not tail-recursive (https://github.com/BuckleScript/bucklescript/blob/80297bbb138f5e0d69c7fed74c5b6043391b2505/lib/js/caml_obj.js#L385).
- It's even worse because in the router, we just want "not equal". If you check the JS output of the diff, you'll see that it's calling `caml_notequal`, which is literally `!caml_notequal`. So we do a full check for deep _equality_ first, _then_ negate it, instead of bailing early in 95% of the cases.
- This is especially bad because ReasonReact is a base library.

So we hand-roll our own `urlNotEqual` comparison:
- It only needs to handle our `url` record. Super lean, and easily and verifiably 100% correct, unlike the generic version.
- The fastest checks possible, with even better readability because it's not generic. See the new JS output; no runtime introspection, no allocation, tail recursive.
- Even better, bails early by checking the other fields first.
- Removes the need for dragging in `caml_obj` (Instagram, and I'm sure some other high-profile highly constrained use-cases, has some informal "be a good citizen" rule regarding introducing new infra; i.e. "don't take over everything, don't be a burden, etc". The Reason community should definitely follow this mantra more, and more so than other FP communities. Don't forcefully shove Reason where you can cause resentments from unrelated teams).

This restores the Router to be mostly allocation-free (edit: except now I see that a rather bad and allocation-heavy implementation of `useUrl` went in... cc @rickyvetter). We can confidently say "this is the fastest and leanest Router for React, not to mention that it actually fulfills more feature requirements of a router than alternatives thanks to our non-opinionated pattern matching on the URL and letting you do the rest (also keep in mind also that OCaml pattern matching doesn't allocate).

Btw, don't get attached to this particular diff's elegant implementation. Generally speaking, we should prefer array over list. It's just that for our Router, list really is a better representation, for once.

Lastly, this diff is one of the reasons why I'm uneasy with all the `deriving` or other codegen for creating equality comparisons. Once you dumbly generate those, you naturally reach for the slowest possible equal comparison (see all the points above), and since most folks don't look at the generated JS, that comparison's cost is entirely hidden now. That's no way to teach performance and correctness.

===

It's hard to write up what "quality" means for Reason libraries, but I'd like this diff to be an example of it. People naively think "quality" means "elegance" and "less visible code", but that's just "cleanness", like how disorganizingly stuffing your clothes into a corner still makes a clean room but doesn't make it easier to find a clothe back. For Reason, I'd like us to mentally inline all the code, pretend they're all visible to you, _then_ evaluate the quality. This has the consequence that we get more careful about dragging in dependencies, that we don't pretend abstracting into a function means the cost is zero now (refer to my recent Reason conf talk), and that we go beyond the superficial polish into deeper polish.